Bluetooth: HID: Reject incoming connection from unpaired device
This patch rejects incoming HID Control Channel request from unpaired
HID devices. Without this patch it was seen that incoming connection
request was accepted for unpaired devices and connection was later
disconnected by DUT when it checked at upper layers that remote device
is not paired, which resulted in hid movements on DUT because of the
delay. Also handling is done for HID Pointing device to be set as
paired by setting linkkey as known when hid connection is up and vice
versa when hid pointing device connection is unpaired.

Change-Id: I4dda40f1f551e04179310e1dfffadf2692bcc5bb
CRs-Fixed: 568386, 568967